requirements.txt
Django==4.2.21
PyMySQL==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0 